pyproject.toml
[tool.poetry]
name = "folding"
version = "0.0.1"
description = "ColabFold modified for AlphaFold2 and local installation"
authors = ["Bevan Hunt"]
readme = "README.md"
license = "MIT, but separate licenses for the trained weights"
repository = "https://github.com/truemagic-coder/cyberchipped-fold"
classifiers = [
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Topic :: Scientific/Engineering :: Artificial Intelligence"
]
packages = [{include = "folding"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
absl-py = "^1.0.0"
jax = { version = "^0.4.20", optional = true }
matplotlib = "^3.2.2"
tensorflow-macos = { version = "<2.14.0", markers = "sys_platform == 'darwin'" }
tensorflow-cpu = { version = "^2.12.1", markers = "sys_platform != 'darwin'" }
biopython = "<1.83"
numpy = "^1.22.0"
pandas = "^1.3.4"
alphafold-colabfold = { version = "v2.3.6", optional = true }
requests = "^2.32.3"
tqdm = "^4.62.2"
appdirs = "^1.4.4"
py3Dmol = "^2.0.1"
dm-haiku = "^0.0.12"
importlib-metadata = "^4.8.2"
torch = "*"


[tool.poetry.dev-dependencies]
ruff = "^0.5.6"
jupyter = "^1.0.0"
pytest = "^8.3.2"
absl-py = "^1.0.0"
dm-haiku = "^0.0.12"
alphafold-colabfold = { version = "v2.3.6" }
mock = "^5.1.0"

[tool.poetry.extras]
alphafold = ["alphafold-colabfold", "jax"]
alphafold-minus-jax = ["alphafold-colabfold"]

[tool.pytest.ini_options]
addopts = "--tb=short"

[tool.poetry.scripts]
folding-batch = 'folding.batch:main'
folding-search = 'folding.mmseqs.search:main'
folding-split_msas = 'folding.mmseqs.split_msas:main'
folding-relax = 'folding.relax:main'

[tool.ruff]
linength = 120
target-version = "py39"
select = ["E", "F", "I", "W"]
ignore = []

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.isort]
known-first-party = ["folding"]

[build-system]
requires = ["poetry-core>=1.0.0,<2.0.0"]
build-backend = "poetry.core.masonry.api"
